requirements.txt
pytorch_lightning==1.9.0  
torchmetrics==0.10.3  
transformers==4.57.3
monai==1.0.1  
pandas==1.4.2  
einops==0.4.1 
open_clip_torch==2.32.0
python==3.10.19
numpy==1.26.4
scipy==1.15.3